requirements.txt
# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0

# Async dependencies with Streamlit fixes
aiohttp>=3.8.0
nest-asyncio>=1.5.0
asyncio-compat>=0.1.2

# Performance optimizations
uvloop>=0.17.0;sys_platform!="win32"  # Linux/Mac only
aiofiles>=23.1.0
